Item List Updated V1.0.2
Item List Update to include Update 4 parts and my predictions on future parts.

Update 4 prediction: When Jace said when showing Assembly Director System it looked like Glados right now Glados works off a Quantum Computer so is this another new part and would it require a new crystal part being created for this? Also another thing thinking about it could we expect to Dark Matter as a part as that is what the part Magnetic Feld Generator would create if you made a a circle with them like the hadron collider? Maybe I reading to much on the science side of this.