Build   : 211220-2217-1021-0006
Ver 1.0.21 Build : 211220-2217-1021-0006

-- Added
ForceClear() -- Some strange reason clearScreen() function is not working as intended so this is a quick fix while I work it out
LayoutMode() -- Can be used to show where the Boarder / Con / Liq / Pwr and system my show on the screen with a # in red. so help finding coordinates on the screen easy.
UVer variable so when we can pull raw data from git / pastebin like in Computercraft the program will be able to say if there is an update for it and then hopefully download from pastebin / git (if there is a way to do this now I have not worked it out yet.

-- Info
Build info consists of Date-Time-Ver-ChangesCount

Changes Count shows how many time it took before I was happy with the outcome the higher the number the more attempts to get it working.

-- General
Testing still in progress.